+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.8.1] - 2026-08-28
+
 ### Fixed
 
 - Station outfitting uploads no longer contain duplicate module entries, which EDDN was accepting but flagging as warnings. Elite lists some modules twice when they can be bought with either credits or Powerplay merc coins.